+++ backend/src/domain/models/__init__.py
@@ -1,0 +1,4 @@
+from .video import Video, VideoPart
+from .user import User
+
+__all__ = ["Video", "VideoPart", "User"]
