+++ backend/src/domain/models/user.py
@@ -1,0 +1,8 @@
+from sqlmodel import SQLModel, Field
+
+
+class User(SQLModel, table=True):
+    __tablename__: str = "users"
+
+    mid: int = Field(primary_key=True)
+    name: str
