+++ backend/src/domain/schemas/__init__.py
@@ -1,0 +1,4 @@
+from .video import VideoSchema, VideoPartSchema
+from .user import UserSchema
+
+__all__ = ["VideoSchema", "VideoPartSchema", "UserSchema"]
