+++ backend/src/domain/schemas/user.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class UserSchema(BaseModel):
+    mid: int
+    name: str
